+++ dbt_semantic_interfaces/dbt_semantic_interfaces/objects/elements/dimension.py
@@ -6,7 +6,7 @@ from dbt_semantic_interfaces.objects.base import HashableBaseModel, ModelWithMet
 from dbt_semantic_interfaces.objects.metadata import Metadata
 from dbt_semantic_interfaces.references import DimensionReference, TimeDimensionReference
 from dbt_semantic_interfaces.enum_extension import ExtendedEnum
-from dbt_semantic_interfaces.objects.time_granularity import TimeGranularity
+from dbt_semantic_interfaces.type_enums.time_granularity import TimeGranularity
 
 ISO8601_FMT = "YYYY-MM-DD"
