+++ dbt_semantic_interfaces/dbt_semantic_interfaces/type_enums/time_granularity.py
