+++ metricflow/constraints/time_constraint.py
@@ -9,7 +9,7 @@ import pandas as pd
 
 from dbt_semantic_interfaces.dataclass_serialization import SerializableDataclass
 from metricflow.time.time_granularity import offset_period
-from dbt_semantic_interfaces.objects.time_granularity import TimeGranularity
+from dbt_semantic_interfaces.type_enums.time_granularity import TimeGranularity
 
 logger = logging.getLogger(__name__)
