+++ metricflow/dataflow/builder/dataflow_plan_builder.py
@@ -68,7 +68,7 @@ from metricflow.specs import (
     InstanceSpecSet,
 )
 from metricflow.sql.sql_plan import SqlJoinType
-from dbt_semantic_interfaces.objects.time_granularity import TimeGranularity
+from dbt_semantic_interfaces.type_enums.time_granularity import TimeGranularity
 
 logger = logging.getLogger(__name__)
