+++ metricflow/dataflow/dataflow_plan.py
@@ -49,7 +49,7 @@ from metricflow.specs import (
     InstanceSpecSet,
 )
 from metricflow.sql.sql_plan import SqlJoinType
-from dbt_semantic_interfaces.objects.time_granularity import TimeGranularity
+from dbt_semantic_interfaces.type_enums.time_granularity import TimeGranularity
 from metricflow.visitor import Visitable, VisitorOutputT
 
 logger = logging.getLogger(__name__)
