+++ metricflow/dataset/convert_data_source.py
@@ -42,7 +42,7 @@ from metricflow.sql.sql_plan import (
     SqlQueryPlanNode,
     SqlSelectQueryFromClauseNode,
 )
-from dbt_semantic_interfaces.objects.time_granularity import TimeGranularity
+from dbt_semantic_interfaces.type_enums.time_granularity import TimeGranularity
 
 logger = logging.getLogger(__name__)
