+++ metricflow/model/semantics/linkable_spec_resolver.py
@@ -24,7 +24,7 @@ from metricflow.specs import (
     EntitySpec,
     EntityReference,
 )
-from dbt_semantic_interfaces.objects.time_granularity import TimeGranularity
+from dbt_semantic_interfaces.type_enums.time_granularity import TimeGranularity
 
 logger = logging.getLogger(__name__)
