+++ metricflow/model/transformations/dbt_to_metricflow.py
@@ -17,7 +17,7 @@ from dbt_semantic_interfaces.objects.metric import Metric, MetricInputMeasure, M
 from dbt_semantic_interfaces.objects.user_configured_model import UserConfiguredModel
 from dbt_semantic_interfaces.parsing.dir_to_model import ModelBuildResult
 from metricflow.model.validations.validator_helpers import ModelValidationResults, ValidationError, ValidationIssue
-from dbt_semantic_interfaces.objects.time_granularity import TimeGranularity
+from dbt_semantic_interfaces.type_enums.time_granularity import TimeGranularity
 
 
 @dataclass
