+++ metricflow/model/validations/dimension_const.py
@@ -14,7 +14,7 @@ from metricflow.model.validations.validator_helpers import (
 )
 from dbt_semantic_interfaces.objects.user_configured_model import UserConfiguredModel
 from dbt_semantic_interfaces.references import DataSourceElementReference, DimensionReference
-from dbt_semantic_interfaces.objects.time_granularity import TimeGranularity
+from dbt_semantic_interfaces.type_enums.time_granularity import TimeGranularity
 
 
 class DimensionConsistencyRule(ModelValidationRule):
