+++ metricflow/naming/linkable_spec_name.py
@@ -4,7 +4,7 @@ import logging
 from dataclasses import dataclass
 from typing import Optional, Tuple
 
-from dbt_semantic_interfaces.objects.time_granularity import TimeGranularity
+from dbt_semantic_interfaces.type_enums.time_granularity import TimeGranularity
 
 DUNDER = "__"
