+++ metricflow/specs.py
@@ -32,7 +32,7 @@ from dbt_semantic_interfaces.dataclass_serialization import SerializableDataclas
 from metricflow.naming.linkable_spec_name import StructuredLinkableSpecName
 from metricflow.sql.sql_bind_parameters import SqlBindParameters
 from metricflow.sql.sql_column_type import SqlColumnType
-from dbt_semantic_interfaces.objects.time_granularity import TimeGranularity
+from dbt_semantic_interfaces.type_enums.time_granularity import TimeGranularity
 
 
 def hash_items(items: Sequence[SqlColumnType]) -> str:
