+++ metricflow/sql/render/big_query.py
@@ -16,7 +16,7 @@ from metricflow.sql.sql_exprs import (
     SqlTimeDeltaExpression,
 )
 from metricflow.sql.sql_plan import SqlSelectColumn
-from dbt_semantic_interfaces.objects.time_granularity import TimeGranularity
+from dbt_semantic_interfaces.type_enums.time_granularity import TimeGranularity
 
 
 class BigQuerySqlExpressionRenderer(DefaultSqlExpressionRenderer):
