+++ metricflow/test/model/validations/test_data_sources.py
@@ -4,7 +4,7 @@ from dbt_semantic_interfaces.objects.data_source import MutabilityType, Mutabili
 from dbt_semantic_interfaces.objects.elements.dimension import Dimension, DimensionType, DimensionTypeParams
 from metricflow.model.validations.validator_helpers import ModelValidationException
 from metricflow.test.model.validations.helpers import data_source_with_guaranteed_meta
-from dbt_semantic_interfaces.objects.time_granularity import TimeGranularity
+from dbt_semantic_interfaces.type_enums.time_granularity import TimeGranularity
 
 
 @pytest.mark.skip("TODO: Will convert to validation rule")
