+++ metricflow/test/model/validations/test_dimension_const.py
@@ -12,7 +12,7 @@ from metricflow.model.validations.data_sources import DataSourceTimeDimensionWar
 from metricflow.model.validations.dimension_const import DimensionConsistencyRule
 from metricflow.model.validations.validator_helpers import ModelValidationException
 from metricflow.test.model.validations.helpers import data_source_with_guaranteed_meta, metric_with_guaranteed_meta
-from dbt_semantic_interfaces.objects.time_granularity import TimeGranularity
+from dbt_semantic_interfaces.type_enums.time_granularity import TimeGranularity
 
 
 def test_incompatible_dimension_type() -> None:  # noqa:D
