+++ metricflow/test/plan_conversion/test_dataflow_to_sql_plan.py
@@ -55,7 +55,7 @@ from metricflow.test.sql.compare_sql_plan import assert_rendered_sql_from_plan_e
 from metricflow.test.sql.compare_sql_plan import assert_sql_plan_text_equal
 from metricflow.test.test_utils import as_datetime
 from metricflow.test.time.metric_time_dimension import MTD_SPEC_DAY
-from dbt_semantic_interfaces.objects.time_granularity import TimeGranularity
+from dbt_semantic_interfaces.type_enums.time_granularity import TimeGranularity
 from dbt_semantic_interfaces.objects.metric import MetricTimeWindow
