+++ metricflow/test/query/test_query_parser.py
@@ -17,7 +17,7 @@ from metricflow.specs import (
 )
 from metricflow.test.test_utils import as_datetime
 from metricflow.test.time.metric_time_dimension import MTD
-from dbt_semantic_interfaces.objects.time_granularity import TimeGranularity
+from dbt_semantic_interfaces.type_enums.time_granularity import TimeGranularity
 from metricflow.time.time_granularity_solver import RequestTimeGranularityException
 from metricflow.test.fixtures.model_fixtures import query_parser_from_yaml
