+++ metricflow/test/sql/test_sql_expr_render.py
@@ -26,7 +26,7 @@ from metricflow.sql.sql_exprs import (
     SqlWindowFunction,
     SqlWindowOrderByArgument,
 )
-from dbt_semantic_interfaces.objects.time_granularity import TimeGranularity
+from dbt_semantic_interfaces.type_enums.time_granularity import TimeGranularity
 
 logger = logging.getLogger(__name__)
