+++ metricflow/test/time/metric_time_dimension.py
@@ -1,5 +1,5 @@
 from metricflow.dataset.dataset import DataSet
-from dbt_semantic_interfaces.objects.time_granularity import TimeGranularity
+from dbt_semantic_interfaces.type_enums.time_granularity import TimeGranularity
 
 # Shortcuts for referring to the metric time dimension.
 MTD = DataSet.metric_time_dimension_name()
