+++ metricflow/time/time_granularity.py
@@ -6,7 +6,7 @@ from typing import Union
 import pandas as pd
 
 from dbt_semantic_interfaces.enum_extension import assert_values_exhausted, ExtendedEnum
-from dbt_semantic_interfaces.objects.time_granularity import TimeGranularity
+from dbt_semantic_interfaces.type_enums.time_granularity import TimeGranularity
 
 
 def offset_period(time_granularity: TimeGranularity) -> pd.offsets.DateOffset:
